// src/users/users.interface.ts
/* eslint-disable prettier/prettier */
/* eslint-disable prettier/prettier */
export interface IUser {
  _id: string;
  name: string;
  email: string;
  role: {
    _id: string;
    name: string;
  };
  permissions?: {
    _id: string;
    name: string;
    apiPath: string;
    module: string;
  };
}